requirements.txt
# Core dependencies
torch>=2.2.0,<2.3.0
neat-python>=0.92
numpy>=1.26.0,<2.0
pandas>=2.3.0
scikit-learn>=1.7.0

# Visualization
matplotlib>=3.3.0
graphviz>=0.16
opencv-python>=4.5.0,<4.12

# Data handling
pmlb>=1.0.0

# Experiment management
jsonschema>=3.2.0
psutil>=5.8.0
click>=8.0.0

# Optional dependencies for extended functionality
GitPython>=3.1.0

# Additional dependencies for compatibility
scipy>=1.16.0

# Database dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0